backend/requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
requests>=2.31.0
